Change encoding of positions to int16_t
uint8_t has been used historically (ayab-firmware) but carriage may move
below needle 0/above 200 (-28:228, i.e. 256/full 8b range) and this choice
is leading to additional math complexity to handle compare/substractions/...
and is therefore error prone.

+++ lib/api/api.h
@@ -106,7 +106,7 @@ class API {
   // Send message to indicate machine state
   void _apiIndicateState(KnitterState state, uint16_t hallValueLeft,
                          uint16_t hallValueRight, CarriageType carriageType,
-                         uint8_t carriagePosition, Direction carriageDirection,
+                         int16_t carriagePosition, Direction carriageDirection,
                          MachineSide hallActive, BeltShift beltshift);
   // Reply to init request message (set machne type)
   void _apiConfirmInit();
